chore(types): move stats types to ~/types/stats and add compatibility re-export
- Add `types/stats.ts` with Stats* interfaces
- Re-export types from `app/repositories/stats.ts` for backward compatibility
- Update imports to use `~/types/stats`
- Adjust RecentRecipe id typing to number
