+++ app/components/FolderNode.vue
@@ -39,7 +39,7 @@
 </template>
 
 <script setup lang="ts">
-import type { Folder } from '~/repositories/folders'
+import type { Folder } from '~/types/folders'
 
 const props = defineProps<{ node: Folder & { children?: Folder[] } }>()
 const emit = defineEmits(['edit', 'delete', 'move'])
