+++ app/pages/folders/index.vue
@@ -71,7 +71,7 @@ import FolderCreateModal from './FolderCreateModal.vue'
 import FolderEditModal from './FolderEditModal.vue'
 import { useFoldersPageStore } from '~/stores/pages/folders'
 import { useFoldersStore } from '~/stores/data/folders'
-import type { Folder } from '~/repositories/folders'
+import type { Folder } from '~/types/folders'
 
 definePageMeta({ requiresAuth: true })
