+++ app/pages/tags/index.vue
@@ -70,7 +70,7 @@
 import { ref, onMounted } from 'vue'
 import { useTagsPageStore } from '~/stores/pages/tags'
 import { useTagsStore } from '~/stores/data/tags'
-import type { Tag } from '~/repositories/tags'
+import type { Tag } from '~/types/tags'
 
 definePageMeta({ requiresAuth: true })
