+++ app/types/stats.ts
